agentic_AI/investment_agent/requirements.txt
anthropic
yfinance
pandas
numpy
pandas-ta
python-dotenv
rich
jinja2
markdown
